[FEAT] Update buildNPCSheets
